requirements.txt
# === Web Application (FastAPI) ===
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2

# === Azure Functions (Durable) ===
azure-functions
azure-functions-durable
azure-storage-queue

# === Azure OpenAI ===
openai
requests==2.32.3
azure-identity

# === Database ===
pyodbc==5.1.0

# === Batch / Discovery (Workflow-2) ===
pandas
numpy
scikit-learn

# === Observability ===
opencensus-ext-azure
pyyaml

# === Testing ===
pytest==8.3.2